+++ game_defence/src/modules/combat_flash_on_damage/components.hpp
@@ -4,8 +4,8 @@ namespace game2d {
 
 struct RequestFlashComponent
 {
-  int milliseconds_left = 0.25 * 1000;
+  int milliseconds_left = static_cast<int>(0.25f * 1000);
   bool flashing = false;
 };
 
-} // namespace game2d
+} // namespace game2d|
